requirements.txt
os
streamlit
gtts
pydub
simpleaudio
speechrecognition
python-dotenv
langchain
langchain_openai
